Move serialization/deserialization functions
Move the consensus serialization and deserialization functions to
`consensus/mod.rs`. This is nice because we can later use
`consensus::serialize` etc. Also for devs opening up the module they see
immediately the most important public functions.

Add re-exports in `encode` and do not change call sites.
