+++ bitcoin/src/consensus/encode.rs
@@ -26,8 +26,10 @@ use crate::blockdata::transaction::{Transaction, TxIn, TxOut};
 // TODO(Tobin): Remove this re-export once consensus refactor is done.
 pub use crate::consensus::decode::{self, Decodable, Error, ReadExt, MAX_VEC_SIZE};
 use crate::consensus::decode::{read_bytes_from_finite_reader, ReadBytesFromFiniteReaderOpts};
+// TODO(Tobin): Remove this re-export once consensus refactor is done.
+pub use crate::consensus::{deserialize, deserialize_partial, serialize, serialize_hex};
 use crate::hash_types::{BlockHash, FilterHash, FilterHeader, TxMerkleNode};
-use crate::io::{self, Cursor};
+use crate::io;
 #[cfg(feature = "std")]
 use crate::p2p::{
     address::{AddrV2Message, Address},
@@ -35,42 +37,6 @@ use crate::p2p::{
 };
 use crate::prelude::*;
 use crate::taproot::TapLeafHash;
-
-/// Encodes an object into a vector.
-pub fn serialize<T: Encodable + ?Sized>(data: &T) -> Vec<u8> {
-    let mut encoder = Vec::new();
-    let len = data.consensus_encode(&mut encoder).expect("in-memory writers don't error");
-    debug_assert_eq!(len, encoder.len());
-    encoder
-}
-
-/// Encodes an object into a hex-encoded string.
-pub fn serialize_hex<T: Encodable + ?Sized>(data: &T) -> String {
-    serialize(data).to_lower_hex_string()
-}
-
-/// Deserializes an object from a vector, will error if said deserialization
-/// doesn't consume the entire vector.
-pub fn deserialize<T: Decodable>(data: &[u8]) -> Result<T, Error> {
-    let (rv, consumed) = deserialize_partial(data)?;
-
-    // Fail if data are not consumed entirely.
-    if consumed == data.len() {
-        Ok(rv)
-    } else {
-        Err(Error::ParseFailed("data not consumed entirely when explicitly deserializing"))
-    }
-}
-
-/// Deserializes an object from a vector, but will not report an error if said deserialization
-/// doesn't consume the entire vector.
-pub fn deserialize_partial<T: Decodable>(data: &[u8]) -> Result<(T, usize), Error> {
-    let mut decoder = Cursor::new(data);
-    let rv = Decodable::consensus_decode_from_finite_reader(&mut decoder)?;
-    let consumed = decoder.position() as usize;
-
-    Ok((rv, consumed))
-}
 
 /// Extensions of `Write` to encode data as per Bitcoin consensus.
 pub trait WriteExt: io::Write {
